refactor(pi-codex-conversion): separate context window budget and wire contracts
