+++ packages/pi-codex-conversion/src/context-management/AGENTS.md
@@ -1,4 +1,5 @@
 - A persisted context-window message is the sole rollover boundary. Pi keeps the full JSONL; provider projection starts at the latest boundary.
+- Window lifecycle stays in `window-manager.ts`; `window-budget.ts` owns reminder deduplication and token budgets, `window-request.ts` owns wire metadata, and `messages.ts` owns persisted message serialization and delivery.
 - Local, Tree and Remote are one context-management contract with 1:1 model-visible behavior. Keep guidance, bootstrap hints, tools, rollover triggers and continuation semantics in parity; only persistence, retrieval, archival and transport may differ. Test shared boundaries across every active mode.
 - Local rollover never invents a summary. Tree archives windows as Pi side branches and filters only owned branch summaries from model context. Remote uses Codex history and notes only on Codex transport and never falls back.
 - Context UUIDs appear in window prompts and turn metadata. Request window IDs use Pi session ID plus zero-based window generation, matching Codex headers.
